Created OC API interface (command-line)
Running oc.py allows you to get either a detailed summary of the upcoming trips for a given stop number, OR just the API's JSON output. You have to pass the App ID and the API Key as parameters.  If you want the latter, throw in the flag '-json' as a third parameter.